# docs/swagger.yaml
definitions:
...
  handlers.TagInput:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  handlers.UpdateProfileInput:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      nickname:
        type: string
      profile_pic:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
...
  response.ProfileResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      nickname:
        type: string
      profile_pic:
        description: Ссылка на фото профиля
        type: string
    type: object
  response.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  response.TagResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  response.TagsListResponse:
    properties:
      tags:
        items:
          $ref: '#/definitions/response.TagResponse'
        type: array
      total:
        type: integer
    type: object
...
  response.UploadAvatarResponse:
    properties:
      message:
        type: string
      profile_pic:
        type: string
    type: object
info:
  contact: {}
  title: '---'
paths:
...
  /profile/delete-avatar:
    delete:
      consumes:
      - application/json
      description: Удаляет аватарку пользователя с сервера и очищает ссылку в базе
        данных.
      produces:
      - application/json
      responses:
        "200":
          description: Аватарка успешно удалена
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "404":
          description: Аватарка не найдена
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление аватарки пользователя
      tags:
      - profile
  /profile/get:
    get:
      consumes:
      - application/json
      description: Получает информацию о пользователе по его ID
      produces:
      - application/json
      responses:
        "200":
          description: Информация о профиле пользователя
          schema:
            $ref: '#/definitions/response.ProfileResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение информации о профиле
      tags:
      - profile
  /profile/update:
    put:
      consumes:
      - application/json
      description: Обновляет информацию профиля пользователя (кроме email)
      parameters:
      - description: Данные для обновления профиля
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProfileInput'
      produces:
      - application/json
      responses:
        "200":
          description: Профиль успешно обновлен
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Ошибка валидации данных
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление информации профиля
      tags:
      - profile
  /profile/upload-avatar:
    post:
      consumes:
      - multipart/form-data
      description: Позволяет пользователю загрузить аватарку. Поддерживаются форматы
        PNG, JPG, JPEG. Максимальный размер файла — 2MB.
      parameters:
      - description: Аватарка пользователя
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Файл успешно загружен
          schema:
            $ref: '#/definitions/response.UploadAvatarResponse'
        "400":
          description: Ошибка валидации (например, файл слишком большой или неподдерживаемый
            формат)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 'Пользователь не найден CODE: USER_NOT_FOUND'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Ошибка сервера (например, ошибка сохранения файла или базы
            данных)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Загрузка аватарки пользователя
      tags:
      - profile
  /tags/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет тег пользователя по id
      parameters:
      - description: ID тега
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Тег успешно удалён
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "404":
          description: Тег не найден TAG_NOT_FOUND
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Ошибка при удалении тега DB_ERROR
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление тега
      tags:
      - tag
  /tags/create:
    post:
      consumes:
      - application/json
      description: Создаёт новый тег
      parameters:
      - description: Данные заметки
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/handlers.TagInput'
      produces:
      - application/json
      responses:
        "201":
          description: Тег успешно создан
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Ошибка при создании тега
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать тег
      tags:
      - tag
  /tags/list:
    get:
      consumes:
      - application/json
      description: Возвращает список тегов пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Список тегов пользователя
          schema:
            items:
              $ref: '#/definitions/response.TagsListResponse'
            type: array
        "500":
          description: Ошибка при получении тегов
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить теги
      tags:
      - tag
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"